refactor: consolidate EntityBrowser and ItemBrowser components
- Renamed ItemBrowser to EntityBrowser for clarity
- Split EntityBrowser into smaller components:
  - EntitiesBrowserTab: handles entity selection
  - CardsBrowserTab: handles card selection (text, separator)
- Removed duplicate code from old EntityBrowser
- Removed unused AddItemButton component
- Added 'Add Items' button when grid is empty
- Updated all imports and tests to reflect new structure
